+++ relay-server/src/actors/server.rs
@@ -5,9 +5,9 @@ use futures::prelude::*;
 use relay_config::Config;
 use relay_statsd::metric;
 
-use crate::actors::controller::{Controller, Shutdown};
 use crate::service;
 use crate::statsd::RelayCounters;
+use relay_common_actors::controller::{Controller, Shutdown};
 
 pub use crate::service::ServerError;